+++ src/nespresso/bot/handlers/client/register.py
@@ -1,9 +1,10 @@
 from aiogram import Dispatcher
 
-from nespresso.bot.handlers.client.commands import start
+from nespresso.bot.handlers.client.commands import find, start
 
 
 def RegisterClientHandlers(dp: Dispatcher) -> None:
     dp.include_routers(
         start.router,
+        find.router,
     )