+++ src/nespresso/bot/lifecycle/menu.py
@@ -5,7 +5,8 @@ from nespresso.bot.lifecycle.creator import bot
 
 async def SetMenu() -> None:
     commands = [
-        BotCommand(command="/start", description="???"),
+        BotCommand(command="/start", description="Register"),
+        BotCommand(command="/find", description="Find alumni by query"),
         BotCommand(command="/cancel", description="Cancel current state"),
         BotCommand(command="/help", description="Help"),
     ]