+++ CHANGELOG.md
@@ -23,7 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.1.0] - 2025-06-04 - Vlad
 ### Added
-- Bot's /start command
+- Client's /start command to register as NES alumni
+- Client's /find command to find alumni by request
 
 ### Updated
 - Keyword extraction